Remove newline in `version.meta` to fix warning
This commit removes the newline character at the end of the single
line of text that is `version.meta`.

Before, running `python3 setup.py install` would cause a `UserWarning`:

```
$ python3 setup.py install
/usr/lib/python3/dist-packages/setuptools/dist.py:285: UserWarning: Normalizing '1.5.6.32
' to '1.5.6.32'
  normalized_version,
running install
```

Note that the error message implicitly indicates the newline character
after "2". The end product of the normalization lacks it.